Split pauli_string_list to separate file and refactor constants to separate file
