+++ src/quaos/core/paulis/constants.py
@@ -1,0 +1,1 @@
+DEFAULT_QUDIT_DIMENSION: int = 2
