+++ src/quaos/core/paulis/pauli_sum.py
@@ -1,6 +1,7 @@
-from typing import Union, overload, Iterable
+from typing import Union, overload
 import numpy as np
 import scipy
+
 from .pauli import Pauli
 from .pauli_string import PauliString
 
@@ -8,70 +9,6 @@ PauliStringDerivedType = Union[list[PauliString], list[Pauli], list[str], PauliS
 PauliType = Union[Pauli, PauliString, 'PauliSum']
 ScalarType = Union[float, complex, int]
 PauliOrScalarType = Union[PauliType, ScalarType]
-
-
-class PauliStringList(list):
-    '''
-    Constructor for PauliStringList class.
-    This class is a wrapper used for tje list of PauliString in a PauliSum.
-    It allows to update the PauliSum tableau any time a PauliString is updated.
-
-    Parameters
-    ----------
-    pauli_list : PauliStringDerivedType
-        The list of PauliStrings representing the operators.
-    weights: list[float | complex] | np.ndarray | None = None
-        The weights for each PauliString.
-    phases: list[int] | np.ndarray | None = None
-        The phases of the PauliStrings in the range [0, lcm(dimensions) - 1].
-    dimensions : list[int] | np.ndarray | int, optional
-        The dimensions of each qudit. If an integer is provided,
-        all qudits are assumed to have the same dimension (default is 2).
-
-    Attributes
-    ----------
-    pauli_list : PauliStringDerivedType
-        The list of PauliStrings representing the operators.
-    weights: list[float | complex] | np.ndarray | None = None
-        The weights for each PauliString.
-    phases: list[int] | np.ndarray | None = None
-        The phases of the PauliStrings in the range [0, lcm(dimensions) - 1].
-    dimensions : list[int] | np.ndarray | int, optional
-        The dimensions of each qudit. If an integer is provided,
-        all qudits are assumed to have the same dimension (default is 2).
-    lcm : int
-        Least common multiplier of all qudit dimensions.
-
-    Raises
-    ------
-    ValueError
-        If the length of pauli_list and weights do not match.
-    '''
-    def __init__(self, data: Iterable["PauliString"], parent: 'PauliSum'):
-        super().__init__(data)
-        self._parent = parent
-
-    @overload
-    def __setitem__(self, key: int, value: 'PauliString') -> None:
-        ...
-
-    @overload
-    def __setitem__(self, key: slice, value: Iterable['PauliString']) -> None:
-        ...
-
-    def __setitem__(self, key, value):
-        """Replace one or more PauliStrings and update the parent's tableau accordingly."""
-        super().__setitem__(key, value)
-
-        if isinstance(key, int):
-            self._parent._tableau[key, :] = self[key].tableau()
-        elif isinstance(key, slice):
-            indices = range(*key.indices(len(self)))
-            for i in indices:
-                self._parent._tableau[i, :] = self[i].tableau()
-
-    # TODO: Add append and extend methods
-    # TODO: Convert to np.ndarray?
 
 
 class PauliSum:
@@ -130,6 +67,7 @@ class PauliSum:
             pauli_list, weights, phases, dimensions
         )
 
+        from .pauli_string_list import PauliStringList
         self.pauli_strings = PauliStringList(sanitized_pauli_list, parent=self)
         self.weights = np.asarray(sanitized_weights, dtype=np.complex128)
         self.dimensions = sanitized_dimensions
