+++ tests/core_tests/paulis_tests/pauli_test.py
@@ -1,6 +1,7 @@
 import numpy as np
 import pytest
 from quaos.core.paulis import PauliSum, PauliString, Pauli
+from quaos.core.paulis.constants import DEFAULT_QUDIT_DIMENSION
 
 
 class TestPaulis:
@@ -195,9 +196,8 @@ class TestPaulis:
 
         assert x1y1 == x1y1_2
 
-        default_dims = PauliString.DEFAULT_QUDIT_DIMENSION
         x1y0 = PauliString([1, 1], [1, 0])
-        x1y0_2 = PauliString.from_string('x1z1 x1z0', dimensions=default_dims)
+        x1y0_2 = PauliString.from_string('x1z1 x1z0', dimensions=DEFAULT_QUDIT_DIMENSION)
         x1y0_3 = PauliString.from_string('x1z1 x1z0', dimensions=dims)
 
         assert x1y0 == x1y0_2
@@ -205,7 +205,7 @@ class TestPaulis:
 
         # Test minimum allowed qudit dimension.
         with pytest.raises(ValueError):
-            _ = PauliString.from_string('x0z0 x0z0', dimensions=PauliString.DEFAULT_QUDIT_DIMENSION - 1)
+            _ = PauliString.from_string('x0z0 x0z0', dimensions=DEFAULT_QUDIT_DIMENSION - 1)
 
     def test_pauli_sum_addition(self):
